Bugfix 1

+++ src/components/GitterPropSetFrom/GitterPropSetFrom.js
@@ -3,11 +3,11 @@ import SetGitterPropSelector from "./SetGitterPropSelector";
 
 class GitterPropSetFrom extends React.Component {
 
-   constructor(props) {
+    constructor(props) {
         super(props);
         this.state = {};
 
-         this.onSubmitEvent = this.onSubmitEvent.bind(this);
+        this.onSubmitEvent = this.onSubmitEvent.bind(this);
     }
 
 
@@ -15,50 +15,53 @@ class GitterPropSetFrom extends React.Component {
     async onSubmitEvent(event) {
         event.preventDefault();
 
-        let temporaryGitterPieceCounter = this.props.appReference.state.gitterPieceCounter;
-        let temporaryMachineCycleTime = this.props.appReference.state.machineCycleTime;
-        let temporaryMachineNestCounter = this.props.appReference.state.machineNestCounter;
-
-        temporaryGitterPieceCounter[parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.piece_count.value);
-        temporaryMachineCycleTime [parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.cycle_time.value);
-        temporaryMachineNestCounter[parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.nest.value);
-
-        let newStateObject = {
-            "gitterPieceCounter": temporaryGitterPieceCounter,
-            "machineCycleTime": temporaryMachineCycleTime,
-            "machineNestCounter": temporaryMachineNestCounter,
-            ...this.props.appReference.state
-        };
-
-        this.props.appReference.setState(newStateObject);
-        localStorage.setItem('gitter_data', JSON.stringify(newStateObject));
+       if (window.confirm(`Konfigurálod a(z) ${this.props.appReference.state.machineNames[parseInt(event.target.gitter_prop_machine_selector.value)]} gépet?`)) {
 
 
+            let temporaryGitterPieceCounter = this.props.appReference.state.gitterPieceCounter;
+            let temporaryMachineCycleTime = this.props.appReference.state.machineCycleTime;
+            let temporaryMachineNestCounter = this.props.appReference.state.machineNestCounter;
+
+            temporaryGitterPieceCounter[parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.piece_count.value);
+            temporaryMachineCycleTime[parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.cycle_time.value);
+            temporaryMachineNestCounter[parseInt(event.target.gitter_prop_machine_selector.value)] = parseInt(event.target.nest.value);
+
+            let newStateObject = {
+                "gitterPieceCounter": temporaryGitterPieceCounter,
+                "machineCycleTime": temporaryMachineCycleTime,
+                "machineNestCounter": temporaryMachineNestCounter,
+                ...this.props.appReference.state
+            };
+
+            this.props.appReference.setState(newStateObject);
+            localStorage.setItem('gitter_data', JSON.stringify(newStateObject));
+      
+        }
     }
 
 
     render() {
-         return (
+        return (
             <>
-            <h4 className="text-center">Gép konfiguráció</h4>
-            <div className="Gitter-control-form-container">
-                <form className="Gitter-control-form row  p-3" onSubmit={this.onSubmitEvent}>
-                    <label className="col-12 mt-2">Kiválasztott gép: <SetGitterPropSelector machines={this.props.appReference.state.machineNames}/></label>
-                    <div className="col-12 mt-2">
-                        
-                        <label className="d-block mt-2">Gitterbe kerülő termékek száma:<input className="form-control" type="number" placeholder="32 (db)" name="piece_count" min="1" required/></label>
+                <h4 className="text-center">Gép konfiguráció</h4>
+                <div className="Gitter-control-form-container">
+                    <form className="Gitter-control-form row  p-3" onSubmit={this.onSubmitEvent}>
+                        <label className="col-12 mt-2">Kiválasztott gép: <SetGitterPropSelector machines={this.props.appReference.state.machineNames} /></label>
+                        <div className="col-12 mt-2">
 
-                        <label className="d-block mt-2">Termék ciklusidje:<input className="form-control" type="number" placeholder="56 (sec)" name="cycle_time" min="1" required/></label>
+                            <label className="d-block mt-2">Gitterbe kerülő termékek száma:<input className="form-control" type="number" placeholder="32 (db)" name="piece_count" min="1" required /></label>
 
-                        <label className="d-block mt-2">Fészekszám:<input className="form-control" type="number" placeholder="2 (db)" name="nest" min="1" required/></label>
+                            <label className="d-block mt-2">Termék ciklusidje:<input className="form-control" type="number" placeholder="56 (sec)" name="cycle_time" min="1" required /></label>
 
-                       
+                            <label className="d-block mt-2">Fészekszám:<input className="form-control" type="number" placeholder="2 (db)" name="nest" min="1" required /></label>
 
-                        <input type="submit" className="w-100 btn btn-outline-secondary GitterControlForm-submit mt-2"
-                               value="Beállít"/>
-                    </div>
-                </form>
-            </div>
+
+
+                            <input type="submit" className="w-100 btn btn-outline-secondary GitterControlForm-submit mt-2"
+                                value="Beállít" />
+                        </div>
+                    </form>
+                </div>
             </>
         );
     }
@@ -66,4 +69,4 @@ class GitterPropSetFrom extends React.Component {
 
 }
 
-export default GitterPropSetFrom ;
+export default GitterPropSetFrom;
